Add files via upload
The page wtas.html is the home page of the website. The CSS styling for this page is main.css
The genre pages are: rock.php, pop.php, love.php, greats.php, edm.php and classical.php. They correspondingly have their own CSS pages.
The sign up and login pages are in register.php and signin.php. The back end for the sign up and login is included in server.php, errors.php and index.php. The CSS page for this is trial.css